chore: TypeScript and small fixes run
