+++ client/src/Pages/Home.tsx
@@ -3,21 +3,19 @@ import { NavLink } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
 
 
-function Home () {
+function Home() {
 
   const navigate = useNavigate();
 
-  useEffect(() => console.log(`http://${process.env.REACT_APP_SERVER_HOST}:${process.env.REACT_APP_SERVER_PORT}`))
-
   return (
     <div id='home-body'>
-     <nav>
-      <ul>
-        <li>
-          <NavLink to="/" className='nav-link'>Tutor Finder Home Page</NavLink>
-        </li>
-      </ul>
-    </nav>
+      <nav>
+        <ul>
+          <li>
+            <NavLink to="/" className='nav-link'>Tutor Finder Home Page</NavLink>
+          </li>
+        </ul>
+      </nav>
       <p>
         TutorFinderTW allows you to find tutors based in Taiwan. Here you can find tutors that teach in person or remotely, or even both! A wide variety of subjects/branches are offered by tutors at different prices, and you can even become one yourself for free by registering!
       </p>
@@ -25,9 +23,8 @@ function Home () {
         <div>
           <button onClick={() => navigate('/register')}>Become a Tutor</button>
         </div>
-
         <div>
-        <button onClick={() => navigate('/tutors')}>Find Tutors</button>
+          <button onClick={() => navigate('/tutors')}>Find Tutors</button>
         </div>
       </div>
     </div>
