+++ client/src/Pages/Tutors.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TutorInterface from '../custom-types/tutor-interface';
+import TutorInterface from '../custom-types/types';
 import TutorCard from '../components/tutor-card';
 import { NavLink } from 'react-router-dom';
 import { useEffect } from 'react';
@@ -10,28 +10,27 @@ type Props = {
 
 
 
-function Tutors ({tutors}: Props) {
+function Tutors({ tutors }: Props) {
 
   useEffect(() => console.log('updated tutors'), [tutors])
 
   return (
     <>
-    <nav>
-      <ul>
-        <li>
-          <NavLink to="/" className='nav-link'>Tutor Finder Home Page</NavLink>
-        </li>
-        <li>
-          <NavLink to="/register" className='nav-link'>Become a Tutor</NavLink>
-        </li>
-      </ul>
-    </nav>
-
+      <nav>
+        <ul>
+          <li>
+            <NavLink to="/" className='nav-link'>Tutor Finder Home Page</NavLink>
+          </li>
+          <li>
+            <NavLink to="/register" className='nav-link'>Become a Tutor</NavLink>
+          </li>
+        </ul>
+      </nav>
       <div id='tutor-cards-wrapper'>
         {
           tutors.map((tutor) => {
             return (
-              <TutorCard key={tutor._id as string} tutorKey={tutor._id as string} tutor={tutor}/>
+              <TutorCard key={tutor._id as string} tutorKey={tutor._id as string} tutor={tutor} />
             );
           })
         }
